fix: decompose AppDelegate and L10n to satisfy SwiftLint file/type length
Split L10n.swift (517 lines) into L10n.swift (72, Common+Service) and
L10n+App.swift (450, App struct + Walkthrough extension).

Extract FloatingPanel, ControlCenterWindow, VisualEffect, and EventMonitor
from AppDelegate.swift (617 lines) into AppDelegate+Windows.swift (132).
Move status menu actions to a private extension in the same file.

All files now under 500-line file_length and 400-line type_body_length
warning thresholds with strict mode enabled.
